Starting a new workflow for myself. Watching a youtube video https://www.youtube.com/watch?v=fju9ii8YsGs I started working through the workshop in Atom and saving each module as a separate file.
After doing a few I though to simultaneously start working with git to increase my everyday use of it. I hope this will also show accountability to myself of what I've learned and credit where from.

My knowledge is improving but not knowing how to connect Atom to the account yet I'm using terminal instead. This may be the wronf place for such verbose notes but I'll pick it up the more I use it hey?